Let the catalogue settle the signatures the headers got wrong
Every C2666 and C2511 here was a header disagreeing with a definition, and
the catalogue says which one is right - so all but two are header fixes and
no proof is traded. `docs/recovery/build-regressions.csv` gains no rows.

C2666 was never an overload-resolution weakness. `has_treaty` is declared
`uint32_t(uint32_t, uint32_t, uint32_t)` in faction.h and DEFINED
`int(int, int, int)` in faction.cpp - two different functions, so every call
was genuinely ambiguous. The original exports `?has_treaty@@YAHHHH@Z`, all
`H`, and AGENTS.md:85 states the rule: parameters are `int`, never
`uint32_t`, because MSVC decorates them `H` and `I` and the wrong one stops
the export matching. The header was the wrong one. Same for
`?get_basic_offense@@YAHHHHHH@Z` and `?get_basic_defense@@YAHHHHH@Z` in
veh.h. 33 errors, three declarations.

C2511 was the same disagreement where it stopped short of ambiguity: five
Path methods, Random::get, Wave_Device::fade and Font::init all had
`uint32_t` in the header against the catalogued `H`. Two went the other way -
`?fade@Sound@@QAEXK@Z` is `unsigned long`, not the `uint32_t` the definition
used, and Wave::load(int, uint32_t) is not catalogued at all so the recovered
definition is the authority and the header follows it. Both of those are
byte-neutral: `unsigned int` and `unsigned long` are the same 32-bit type
passed the same way, so only the decorated name moves.

The remaining C++11: dialogs_recovery.cpp had a lambda, which becomes the
plain function it always was with its one capture passed explicitly; VC6 has
no lambdas. Bare `{}` inside aggregate initialisers becomes `{0}`, and
`ScrollCloseButtonTrace{}` becomes `ScrollCloseButtonTrace()`.

tools/generate_vc6_exports.py is new and not yet wired. The MinGW export
generator exists to BRIDGE two manglings - it builds the DLL, reads its
Itanium symbols back through nm and pairs them to MSVC imports by signature,
which is why every line it writes is an alias. VC6 needs none of that: it
emits the MSVC name directly, so the export list is just the patched
executable's import list and no build has to exist first. Left unwired
deliberately - the checked-in .def carries 488 names and both patched images
present here import 462, a strict subset, so pinning the larger list today
would invent 26 unresolved externals. Nothing is lost by waiting: under MSVC
the exports come from `__declspec(dllexport)`, which the headers already
carry (`class DLLEXPORT Heap`), so the export list does not vanish without a
.def the way it would under MinGW.

+++ src/sound.cpp
@@ -41,7 +41,7 @@ Original Offset: 004C6600
 Return Value: n/a
 Status: Complete
 */
-void Sound::fade(uint32_t a1) {
+void Sound::fade(unsigned long a1) {
     // The object's own vtable is read at run time rather than declaring these
     // virtual, so the dispatch cannot disagree with the original's layout.
     uint8_t *const vtable = *reinterpret_cast<uint8_t **>(this);
